requirements.txt
# IO Board Communication Library - Dependencies

# Serial communication
pyserial>=3.5

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Code quality
black>=23.0
mypy>=1.0

# Type stubs for serial (optional)
types-pyserial>=3.5
